+++ lib/commons.dart
@@ -15,10 +15,14 @@ export 'package:commons/src/functions/navigation_functions.dart';
 export 'package:commons/src/shared_preferences/shared_prefrences.dart';
 export 'package:commons/src/toasts/toasts.dart';
 export 'package:commons/src/screens/loading_screen.dart';
+export 'package:commons/src/screens/widget_view_screen.dart';
+export 'package:commons/src/validators/validator.dart';
+export 'package:commons/src/functions/try_catch.dart';
 
 // Dependencies export
 export 'package:url_launcher/url_launcher.dart';
 export 'package:device_info/device_info.dart';
+export 'package:package_info/package_info.dart';
 export 'package:share/share.dart';
 export 'package:shared_preferences/shared_preferences.dart';
 export 'package:oktoast/oktoast.dart';